0.27.0: get rid of async_trait
It proved to be counter-productive to use async_trait, because
it costs us another allocation, makes the interface more complicated
and makes it extremely painful to make clients and future
Send. Because of that, a breaking API change renames the GenericClient
to just Client, and from now on, it should just be the entrypoint
for using the driver. It's a simple enum over multiple backends,
and it will automatically and transiently works for all of them,
while also being Send, and making its futures Send as well.